requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
orjson==3.11.4
propcache==0.4.1
redis==7.1.0
typing_extensions==4.15.0
yarl==1.22.0
yt-dlp==2025.11.12
